requirements.txt
streamlit_folium
psutil
localtileserver
streamlit
rasterio
geemap
leafmap
streamlit-image-comparison
# For Copernicus
python-dotenv
sentinelhub
geopy
shapely
folium
pandas
numpy==1.24.2
click
python-dotenv
openai
joblib==1.2.0
scikit-learn==1.2.2
requests
